Fix Windows portability diagnostics

+++ src/utils/process.rs
@@ -45,7 +45,7 @@ pub(crate) fn is_process_alive(pid: u32) -> bool {
         let mut exit_code: u32 = 0;
         let ok = GetExitCodeProcess(handle, &mut exit_code);
         CloseHandle(handle);
-        ok != 0 && exit_code == STILL_ACTIVE
+        ok != 0 && exit_code == STILL_ACTIVE as u32
     }
 }
